requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
streamlit==1.45.1
python-pptx==0.6.23
python-dotenv==1.0.0
supabase==2.8.0
openai==1.54.4
lxml==4.9.3
bcrypt==4.1.2